Updated all classes and models representing money to use decimal (there was an inconsistency with the conf. managment)

+++ source/Conference/Payments/ThirdPartyProcessorPayment.cs
@@ -32,7 +32,7 @@ namespace Payments
 
         private List<IEvent> events = new List<IEvent>();
 
-        public ThirdPartyProcessorPayment(Guid id, Guid paymentSourceId, string description, double totalAmount, IEnumerable<ThidPartyProcessorPaymentItem> items)
+        public ThirdPartyProcessorPayment(Guid id, Guid paymentSourceId, string description, decimal totalAmount, IEnumerable<ThidPartyProcessorPaymentItem> items)
             : this()
         {
             this.Id = id;
@@ -69,7 +69,7 @@ namespace Payments
 
         public string Description { get; private set; }
 
-        public double TotalAmount { get; private set; }
+        public decimal TotalAmount { get; private set; }
 
         public virtual ICollection<ThidPartyProcessorPaymentItem> Items { get; private set; }
